feat: Split APIs by permission level
