+++ .generated-sources/emily/openapi/api-config/admin.json
@@ -1,0 +1,6 @@
+{
+  "packageName": "admin-emily-client",
+  "packageVersion": "0.1.0",
+  "packageDescription": "Client for the Emily api admin endpoint (generated by OpenAPI Generator)",
+  "preferUnsignedInt": true
+}
