+++ .generated-sources/emily/openapi/api-config/public.json
@@ -1,0 +1,6 @@
+{
+  "packageName": "emily-client",
+  "packageVersion": "0.1.0",
+  "packageDescription": "Client for the Emily api public endpoint (generated by OpenAPI Generator)",
+  "preferUnsignedInt": true
+}
