+++ .generated-sources/emily/openapi/api-config/signer.json
@@ -1,0 +1,6 @@
+{
+  "packageName": "signer-emily-client",
+  "packageVersion": "0.1.0",
+  "packageDescription": "Client for the Emily api signer endpoint (generated by OpenAPI Generator)",
+  "preferUnsignedInt": true
+}
