+++ .generated-sources/emily/openapi/api-config/testing.json
@@ -1,0 +1,6 @@
+{
+  "packageName": "testing-emily-client",
+  "packageVersion": "0.1.0",
+  "packageDescription": "Client for the Emily api testing endpoint (generated by OpenAPI Generator)",
+  "preferUnsignedInt": true
+}
