--- .generated-sources/emily/openapi/config.json
@@ -1,5 +1,0 @@
-{
-    "packageName": "emily-client",
-    "packageVersion": "0.1.0",
-    "packageDescription": "Client for the Emily api (generated by OpenAPI Generator)"
-  }
